Add BMI to person detail page

+++ MvcMoviesCore/ApiController/PersonApiController.cs
@@ -597,7 +597,7 @@ namespace MvcMoviesCore.ApiController
             }
 
             persons.ToList().ForEach(f => f.ActorsAge = f.GetActorsAge(f.Birthday, f.Obit));
-            persons.ToList().ForEach(f => f.Value = f.GetBMI(f.Height, f.Weight));
+            persons.ToList().ForEach(f => f.Value = (int)f.GetBMI(f.Height, f.Weight));
             foreach (var person in persons)
             {
                 person.MoviesPerson.ToList().ForEach(f => f.Person = null);
